# crm-it-shkola-rtk/backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/users": {
      "get": {
        "operationId": "AuthController_getUsers",
        "parameters": [
          {
            "name": "X-Dev-Role",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "example": "kam"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserDto"
                  }
                }
              }
            }
          }
        },
        "summary": "Справочник сотрудников (ФИО, роль, руководитель) — только чтение",
        "tags": [
          "auth"
        ]
      }
    },
    "/admin/users": {
      "get": {
        "operationId": "AdminController_getUsers",
        "parameters": [
          {
            "name": "X-Dev-Role",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "example": "administrator"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserDto"
                  }
                }
              }
            }
          }
        },
        "summary": "Пользователи и их роли",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/users/{id}": {
      "patch": {
        "operationId": "AdminController_patchUpdateUser",
        "parameters": [
          {
            "name": "X-Dev-Role",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "example": "administrator"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "example": "c0000000-0000-4000-8000-000000000001",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          }
        },
        "summary": "Изменить роль/активность/руководителя пользователя",
        "tags": [
          "admin"
        ]
      },
      "put": {
        "deprecated": true,
        "operationId": "AdminController_updateUser",
        "parameters": [
          {
            "name": "X-Dev-Role",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "example": "administrator"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "example": "c0000000-0000-4000-8000-000000000001",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          }
        },
        "summary": "Изменить роль/активность/руководителя пользователя",
        "tags": [
          "admin"
        ]
      }
    },
...
  },
  "info": {
    "title": "CRM ИТ Школа РТК",
...
    "version": "0.0.1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "UserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "c0000000-0000-4000-8000-000000000001"
          },
          "email": {
            "type": "string",
            "example": "kam@it-shkola-rtk.ru"
          },
          "fullName": {
            "type": "string",
            "example": "Иванова Мария Сергеевна"
          },
          "role": {
            "enum": [
              "KAM",
              "RUKOVODITEL",
              "ADMINISTRATOR"
            ],
            "type": "string",
            "example": "KAM"
          },
          "isActive": {
            "type": "boolean",
            "example": true
          },
          "managerId": {
            "type": "string",
            "nullable": true,
            "example": "c0000000-0000-4000-8000-000000000002"
          }
        },
        "required": [
          "id",
          "email",
          "fullName",
          "role",
          "isActive"
        ]
      },
...
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "role": {
            "enum": [
              "KAM",
              "RUKOVODITEL",
              "ADMINISTRATOR"
            ],
            "type": "string",
            "example": "RUKOVODITEL"
          },
          "isActive": {
            "type": "boolean",
            "example": false
          },
          "managerId": {
            "type": "string",
            "nullable": true,
            "example": "c0000000-0000-4000-8000-000000000002"
          }
        }
      },
...
    }
  }
}